requirements.txt
livekit-agents
livekit-plugins-elevenlabs
python-dotenv 